Move data generators to nupic.data.generators module
