+++ nupic/data/generators/distributions.py
